Feat: finalização do projeto

+++ src/main/ecommerce/core/domain/contract/desconto/DescontoRepository.java
@@ -1,5 +1,5 @@
 package main.ecommerce.core.domain.contract.desconto;
 
 public interface DescontoRepository {
-    public double aplicarDesconto (double total);
+    public float aplicarDesconto (float total);
 }
