+++ src/main/ecommerce/core/domain/service/PrimeiraCompra.java
@@ -5,7 +5,7 @@ import main.ecommerce.core.domain.contract.desconto.DescontoRepository;
 public class PrimeiraCompra implements DescontoRepository {
 
     @Override
-    public double aplicarDesconto(double total) {
-        return total - (total * (50.0 / 100.0));
+    public float aplicarDesconto(float total) {
+        return total - (total * (50 / 100));
     }
 }
